Pass config to onnx scripts
Insted of trying to import config in every file, pass it around as an
argument. This helps us to load the same config in all files.
Now we can also specify some different config to be used. This allows
us to run onnx generation in pipeline with test config instead of
automatically importing the failnet config in the same directory.
Now we are not forced to keep the "cpu" setting in failnet/config.py